Fix string lists to have a sensible default value handling.
Previously, if you set a default value for a string list
setting, the user's settings were appended to it. Now the
default is used if the user does not specify the setting
at all (same as before), but ignored if the user does
specify. Thus, if the default for foo was ['bar'], then
--foo=pink now results in ['pink'], rather than
['bar', 'pink'].