Fix default value handling for string list settings.
Previously, --help would always show [] for the string list
default value, even if the real default was something else.
Also, there was a bad interaction between non-empty default
values and config files.

Now, --help shows the current value (default or from config
file), same as for other setting types, and string list
values in config files work even if the default is non-empty.

Thanks to this, the whole default_value property business
of Setting objects was removed, which is a relief, since it
didn't really do anything useful. This is, however, an API
change, but I doubt anyone was using this part of it.