Fix cliapp.runcmd to not die from SIGWINCH
If the application is reacting to SIGWINCH, for example by using ttystatus,
it will get that signal whenever the terminal window size changes. This causes
select.select() to raise an exception, which killed the program. We now
catch the exception, check that the exception is for an interrupted system
call (EINTR), and if so, pretend as if nothing happened. Any other reasons
for the exception are re-raised.