runcmd: Collect stderr of whole pipeline
When pipe_stderr is None (inherit parent fd), subprocess.STDOUT (stderr to
stdout) or a specific file, it would put everything to the right place, but
when pipe_stderr is subprocess.PIPE the behaviour is less well defined.

Before this change, it would happen to return the stderr of the last element of
the pipeline, behaving similar to this shell snippet, except it does not need
an intermediate file to store the contents of stderr in memory.

    stdout="$(foo 2>/dev/null | bar 2>/dev/null | baz 2>tmp)"
    stderr="$(cat tmp)"

With this patch, it now behaves more like:

    stdout="$((foo | bar | baz) 2>tmp)"
    stderr="$(cat tmp)"

It works by duplicating some of the subprocess.PIPE behaviour outside of
subprocess.Popen, as there's no better way to make every subprocess have the
same stderr, since there's no way to get the write end out of the Popen object
after it has been constructed, since it is closed in the constructor, only
leaving the read-end available as p.stderr.

Only the last element of the pipeline is given the read end of the pipe, since
it ought to have only one read pipe, and it makes the most sense for the last
element to have the pipe, since you can then think of the pipeline as having a
pipe that joins every element for stderr, that comes out at the end, with the
stdout of the pipeline.